fix: change Price type float to decimal

+++ WonderfulStore.Core/Entities/Product.cs
@@ -2,7 +2,7 @@ namespace WonderfulStore.Core.Entities
 {
     public class Product : IBaseEntity
     {
-        public Product(string name, float price, string description, Guid? idPromotion)
+        public Product(string name, decimal price, string description, Guid? idPromotion)
         {
             Id = Guid.NewGuid();
             Name = name;
@@ -16,7 +16,7 @@ namespace WonderfulStore.Core.Entities
 
         public Guid Id {get; private set;}
         public string Name { get; private set; }
-        public float Price { get; private set; }
+        public decimal Price { get; private set; }
         public string Description { get; private set; }
         public Guid? IdPromotion { get; private set; }
         public Promotion? Promotion { get; private set; }
